CORDA-3377: Make API more Java-friendly and resolve some simple warnings.

+++ djvm/src/main/java/sandbox/java/lang/Integer.java
@@ -66,7 +66,7 @@ public final class Integer extends Number implements Comparable<Integer> {
 
     @Override
     public double doubleValue() {
-        return (double) value;
+        return value;
     }
 
     @Override
